requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
aiohttp-apispec==2.2.1
aiohttp-jinja2==1.4.2
aiohttp-middlewares==1.1.0
aiohttp-swagger==1.0.15
aioredis==1.3.1
aioreloader==0.3.0
alembic==1.6.5
apispec==3.3.2
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
black==21.7b0
brotlipy==0.7.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
flake8==3.9.2
greenlet==1.1.0
hiredis==2.0.0
idna==3.2
isort==5.9.2
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.13.0
mccabe==0.6.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==5.3.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
tomli==1.0.4
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvloop==0.15.2
webargs==5.5.3
yarl==1.6.3
